requirements.txt
PyGithub==2.1.1
selenium==4.18.1
webdriver-manager==4.0.1
Pillow==10.2.0
python-dotenv==1.0.1 